requirements.txt
# Core GUI Framework
PyQt5>=5.15.0
PyQtWebEngine>=5.15.0

# Document Processing
pdfplumber>=0.9.0
python-docx>=0.8.11
openpyxl>=3.1.0
pandas>=1.5.0
PyMuPDF>=1.23.0

# OCR Support (Optional)
pytesseract>=0.3.10
pdf2image>=1.16.0
Pillow>=9.0.0

# GIS and Geospatial
geopandas>=0.12.0
pyproj>=3.4.0
shapely>=1.8.0

# LLM Integration
requests>=2.28.0
openai>=1.0.0

# Data Export
xlsxwriter>=3.0.0

# ArcGIS Integration (if available)
# arcpy (comes with ArcGIS Pro installation)

# Development and Testing
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0 